requirements.txt
django>=1.5.1

# Additional requirements go here
django-floppyforms>=1.1
django-crispy-forms>=1.4.0